% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping.R
\name{ggsave}
\alias{ggsave}
\title{Wrapper for ggsave}
\usage{
ggsave(
  filename,
  plot = last_plot(),
  device = NULL,
  path = NULL,
  scale = 1,
  width = 3.5,
  height = 2.5,
  units = "in",
  dpi = "retina",
  limitsize = TRUE,
  ...
)
}
\arguments{
\item{filename}{see documentation of \link[ggplot2]{ggsave}}

\item{plot}{see documentation of \link[ggplot2]{ggsave}}

\item{device}{see documentation of \link[ggplot2]{ggsave}}

\item{path}{see documentation of \link[ggplot2]{ggsave}}

\item{scale}{see documentation of \link[ggplot2]{ggsave}}

\item{width}{see documentation of \link[ggplot2]{ggsave}}

\item{height}{see documentation of \link[ggplot2]{ggsave}}

\item{units}{see documentation of \link[ggplot2]{ggsave}}

\item{dpi}{see documentation of \link[ggplot2]{ggsave}}

\item{limitsize}{see documentation of \link[ggplot2]{ggsave}}

\item{...}{see documentation of \link[ggplot2]{ggsave}}
}
\value{
see documentation of \link[ggplot2]{ggsave}
}
\description{
This function wraps around the ggsave function of gglpot2 (\link[ggplot2]{ggsave}) by setting the defaults to the SGB standard output.
}
\examples{
ggsave("my_sgb_plot.png")
}
