% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{ggtheme}
\alias{theme_sgb}
\alias{theme_sgb_blank}
\title{Standard SGB-Themes fuer ggplot2}
\usage{
theme_sgb(
  base_size = 8,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_sgb_blank(
  base_size = 9,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Standard Schriftgroesse (optional, numeric).
Default: \code{base_size = 8}.}

\item{base_family}{Standard Schriftart (optional, character).
Default: \code{base_family = ""}.}

\item{base_line_size}{Standard Liniengroesse (optional, numeric).
Default: \code{base_line_size = base_size/22}.}

\item{base_rect_size}{Standard Groesse rechteckige Elemente (optional, numeric).
Default: \code{base_rect_size = base_size/22}.}
}
\description{
Das Theme basiert auf dem ggplot2 theme_gray und theme_classic. Je nach verwendeter Grafik
muessen einzelne Elemente selber noch angepasst werden (z.B. Position der Legende oder
Entfernen der y-Achsenbeschriftung).
Zuesaetzlich wird auch ein minimalistisches Theme definiert, etwa fuer Kuchendiagramme.
}
\details{
\describe{

\item{`theme_sgb`}{
Das Standard SGB-Theme, mit weissem Hintergrund, schwarzer x- und y-Achse, ohne
Rahmen um den Plot und grauen Rasterlinien.}

\item{`theme_sgb_blank`}{
Minimalistisches Theme basierend auf theme_sgb. Ohne Achsen, Raster, Beschriftungen etc.
Geeignet etwa fuer Kuchendiagramme.
}
}
}
\examples{

\donttest{
  # Plotten mit iris Daten (mit ggplot2, theme_sgb, and SGB Farben):

  library("ggplot2")  # theme_sgb benoetigt ggplot2

  ggplot(datasets::iris) +
     geom_point(aes(Sepal.Width, Sepal.Length, color = Species),
       size = 2) +
     scale_color_manual(values = usecol(pal = pal_sgb_pref, n = 3)) +
     theme_sgb(base_size = 10) +
     theme(axis.title.y = element_blank()) # manually remove y-Axis title
}

}
\seealso{
Other plot functions: 
\code{\link{ggexample}}
}
\concept{plot functions}
