% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kof}
\alias{kof}
\title{KOF-Beispieldaten}
\format{
Ein Dataframe mit 591 Zeilen und 4 Variablen.
}
\usage{
kof
}
\description{
Der Datensatz enthaelt eine Variable aus den KOF-Konjunkturdaten fuer drei Branchen.
Die Daten sind im long-Format, sodass sie direkt mit ggplot verwendet werden koennen.
}
\keyword{datasets}
