% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mangelindikator}
\alias{mangelindikator}
\title{Mangelindikator Beispieldaten}
\format{
Ein Dataframe mit 260 Zeilen und 4 Variablen.
}
\usage{
mangelindikator
}
\description{
Der Datensatz enthaelt den Mangelindikator fuer eine Branche, mit vier Ausbildungslevels.
Die Daten sind im long-Format, sodass sie direkt mit ggplot verwendet werden koennen.
}
\keyword{datasets}
