% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def.R
\docType{data}
\name{pal_sgb}
\alias{pal_sgb}
\title{SGB Standard-Farbpalette (11 colors).}
\format{
An object of class \code{character} of length 11.
}
\usage{
pal_sgb
}
\description{
\code{pal_sgb} kombiniert die 5 roten Farben der
Farbpalette \code{\link{pal_sgb_rot}}
mit den 6 nicht-roten Farben von \code{\link{pal_sgb}}
zu einer Palette mit 11 Farbwerten. Damit ist \code{weiss}
in der Mitte der Palette. Das ist hilfreich, wenn Gradienten
verwendet werden.
}
\examples{
pal_sgb

# Auswahl nach Position
pal_sgb[1]    # Farbname "rot5" (as df)
pal_sgb[[1]]  # Farbwert "#660000"

# Auswahl nach Name
pal_sgb["rot5"]   # Farbname "rot5" (as df)
pal_sgb[["rot5"]] # Farbwert "#660000"

# Palette plotten:
seecol(pal_sgb)

}
\seealso{
\code{\link{pal_sgb}} fuer die SGB Standard-Farbpalette mit allen 5 Farben von \code{\link{pal_sgb_rot}};
\code{\link{pal_sgb_rot}} fuer die rote SGB-Farbpalette;
\code{\link{seecol}} um die Farbpaletten anzusehen;
\code{\link{usecol}} um die Farbpaletten zu verwenden.

Other color palettes: 
\code{\link{pal_sgb_dunkelblau}},
\code{\link{pal_sgb_gelb}},
\code{\link{pal_sgb_grau}},
\code{\link{pal_sgb_hellblau}},
\code{\link{pal_sgb_pair}},
\code{\link{pal_sgb_pref}},
\code{\link{pal_sgb_rot}},
\code{\link{pal_sgb_signal}},
\code{\link{pal_sgb_web}}
}
\concept{color palettes}
\keyword{datasets}
