% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_1.R
\docType{data}
\name{pal_sgb_dunkelblau}
\alias{pal_sgb_dunkelblau}
\title{Dunkelblaue SGB-Palette}
\format{
An object of class \code{data.frame} with 1 rows and 5 columns.
}
\usage{
pal_sgb_dunkelblau
}
\description{
\code{pal_sgb_dunkelblau} ist eine zusaetzliche SGB-Farbpalette als data frame mit 5 Farben
(Abstufungen von \code{\link{SGBdunkelblau}}).
}
\examples{
pal_sgb_dunkelblau

# Auswahl nach Position:
pal_sgb_dunkelblau[4]    # 2. Farbname "dunkelblau4" (als df)
pal_sgb_dunkelblau[[4]]  # 2. Farbwert "#42467e"

# Auswahl nach Name:
pal_sgb_dunkelblau["dunkelblau4"]    # Farbe "dunkelblau4" (als df)
pal_sgb_dunkelblau[["dunkelblau4"]]  # Farbwert "#42467e"

# Palette plotten:
seecol(pal_sgb_dunkelblau)

}
\seealso{
\code{\link{pal_sgb}} fuer die SGB Standard-Farbpalette mit allen 5 Farben von \code{\link{pal_sgb_rot}};
\code{\link{pal_sgb_pref}} fuer eine SGB Farbpalette mit allen bevorzugten Farben;
\code{\link{seecol}} um die Farbpaletten anzusehen;
\code{\link{usecol}} um die Farbpaletten zu verwenden.

Other color palettes: 
\code{\link{pal_sgb_gelb}},
\code{\link{pal_sgb_grau}},
\code{\link{pal_sgb_hellblau}},
\code{\link{pal_sgb_pair}},
\code{\link{pal_sgb_pref}},
\code{\link{pal_sgb_rot}},
\code{\link{pal_sgb_signal}},
\code{\link{pal_sgb_web}},
\code{\link{pal_sgb}}
}
\concept{color palettes}
\keyword{datasets}
