% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_1.R
\docType{data}
\name{pal_sgb_gelb}
\alias{pal_sgb_gelb}
\title{Gelbe SGB-Palette}
\format{
An object of class \code{data.frame} with 1 rows and 5 columns.
}
\usage{
pal_sgb_gelb
}
\description{
\code{pal_sgb_gelb} ist eine zusaetzliche SGB-Farbpalette als data frame mit 5 Farben
(Abstufungen von \code{\link{SGBgelb}}).
}
\examples{
pal_sgb_gelb

# Auswahl nach Position:
pal_sgb_gelb[4]    # 2. Farbname "gelb4" (als df)
pal_sgb_gelb[[4]]  # 2. Farbwert "#ffce58"

# Auswahl nach Name:
pal_sgb_gelb["gelb4"]    # Farbe "gelb4" (als df)
pal_sgb_gelb[["gelb4"]]  # Farbwert "#ffce58"

# Palette plotten:
seecol(pal_sgb_gelb)

}
\seealso{
\code{\link{pal_sgb}} fuer die SGB Standard-Farbpalette mit allen 5 Farben von \code{\link{pal_sgb_rot}};
\code{\link{pal_sgb_pref}} fuer eine SGB Farbpalette mit allen bevorzugten Farben;
\code{\link{seecol}} um die Farbpaletten anzusehen;
\code{\link{usecol}} um die Farbpaletten zu verwenden.

Other color palettes: 
\code{\link{pal_sgb_dunkelblau}},
\code{\link{pal_sgb_grau}},
\code{\link{pal_sgb_hellblau}},
\code{\link{pal_sgb_pair}},
\code{\link{pal_sgb_pref}},
\code{\link{pal_sgb_rot}},
\code{\link{pal_sgb_signal}},
\code{\link{pal_sgb_web}},
\code{\link{pal_sgb}}
}
\concept{color palettes}
\keyword{datasets}
