% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def.R
\docType{data}
\name{pal_sgb_hellblau}
\alias{pal_sgb_hellblau}
\title{Hellblaue SGB-Palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
pal_sgb_hellblau
}
\description{
\code{pal_sgb_hellblau} ist eine zusaetzliche SGB-Farbpalette als data frame mit 5 Farben
(Abstufungen von \code{\link{SGBhellblau}}).
}
\examples{
pal_sgb_hellblau

# Auswahl nach Position:
pal_sgb_hellblau[4]    # 2. Farbname "hellblau4" (als df)
pal_sgb_hellblau[[4]]  # 2. Farbwert "#63bae6"

# Auswahl nach Name:
pal_sgb_hellblau["hellblau4"]    # Farbe "hellblau4" (als df)
pal_sgb_hellblau[["hellblau4"]]  # Farbwert "#63bae6"

# Palette plotten:
seecol(pal_sgb_hellblau)

}
\seealso{
\code{\link{pal_sgb}} fuer die SGB Standard-Farbpalette mit allen 5 Farben von \code{\link{pal_sgb_rot}};
\code{\link{pal_sgb_pref}} fuer eine SGB Farbpalette mit allen bevorzugten Farben;
\code{\link{seecol}} um die Farbpaletten anzusehen;
\code{\link{usecol}} um die Farbpaletten zu verwenden.

Other color palettes: 
\code{\link{pal_sgb_dunkelblau}},
\code{\link{pal_sgb_gelb}},
\code{\link{pal_sgb_grau}},
\code{\link{pal_sgb_pair}},
\code{\link{pal_sgb_pref}},
\code{\link{pal_sgb_rot}},
\code{\link{pal_sgb_signal}},
\code{\link{pal_sgb_web}},
\code{\link{pal_sgb}}
}
\concept{color palettes}
\keyword{datasets}
