% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_2.R
\docType{data}
\name{pal_sgb_pair}
\alias{pal_sgb_pair}
\title{Paarweise SGB-Farben in einer Farbpalette}
\format{
An object of class \code{data.frame} with 1 rows and 10 columns.
}
\usage{
pal_sgb_pair
}
\description{
\code{pal_sgb_pref} ist eine zusaetzliche SGB-Farbpalette, welche die
5 SGB-Farben (rot, hellblau, dunkelblau, gelb, grau) in 10 Paaren zusammenstellt.
}
\examples{
pal_sgb_pair

# Palette plotten:
seecol(pal_sgb_pair)

}
\seealso{
\code{\link{pal_sgb}} fuer die SGB Standard-Farbpalette mit allen 5 Farben von \code{\link{pal_sgb_rot}};
\code{\link{pal_sgb_rot}} fuer die rote SGB-Farbpalette;
\code{\link{seecol}} um die Farbpaletten anzusehen;
\code{\link{usecol}} um die Farbpaletten zu verwenden.

Other color palettes: 
\code{\link{pal_sgb_dunkelblau}},
\code{\link{pal_sgb_gelb}},
\code{\link{pal_sgb_grau}},
\code{\link{pal_sgb_hellblau}},
\code{\link{pal_sgb_pref}},
\code{\link{pal_sgb_rot}},
\code{\link{pal_sgb_signal}},
\code{\link{pal_sgb_web}},
\code{\link{pal_sgb}}
}
\concept{color palettes}
\keyword{datasets}
