% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def.R
\docType{data}
\name{pal_sgb_pref}
\alias{pal_sgb_pref}
\title{Bevorzugte SGB-Farben in einer Farbpalette}
\format{
An object of class \code{character} of length 5.
}
\usage{
pal_sgb_pref
}
\description{
\code{pal_sgb_pref} ist eine zusaetzliche SGB-Farbpalette
mit allen bevorzugten SGB-Farben (rot, hellblau, dunkelblau, gelb, grau).
}
\examples{
pal_sgb_pref

# Palette plotten:
seecol(pal_sgb_pref)

}
\seealso{
\code{\link{pal_sgb}} fuer die SGB Standard-Farbpalette mit allen 5 Farben von \code{\link{pal_sgb_rot}};
\code{\link{seecol}} um die Farbpaletten anzusehen;
\code{\link{usecol}} um die Farbpaletten zu verwenden.

Other color palettes: 
\code{\link{pal_sgb_dunkelblau}},
\code{\link{pal_sgb_gelb}},
\code{\link{pal_sgb_grau}},
\code{\link{pal_sgb_hellblau}},
\code{\link{pal_sgb_pair}},
\code{\link{pal_sgb_rot}},
\code{\link{pal_sgb_signal}},
\code{\link{pal_sgb_web}},
\code{\link{pal_sgb}}
}
\concept{color palettes}
\keyword{datasets}
