% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def.R
\docType{data}
\name{pal_sgb_rot}
\alias{pal_sgb_rot}
\title{Rote SGB-Farbpalette}
\format{
An object of class \code{character} of length 5.
}
\usage{
pal_sgb_rot
}
\description{
\code{pal_sgb_rot} ist eine zusaetzliche SGB Farbpalette als data frame mit 5 Farben
(Abstufungen von \code{\link{SGBrot}}).
}
\examples{
pal_sgb_rot

# Auswahl nach Position:
pal_sgb_rot[4]    # 2. Farbname "rot4" (als df)
pal_sgb_rot[[4]]  # 2. Farbwert "#990000"

# Auswahl nach Name:
pal_sgb_rot["rot4"]    # Farbe "rot4" (als df)
pal_sgb_rot[["rot4"]]  # Farbwert "#990000"

# Palette plotten:
seecol(pal_sgb_rot)

}
\seealso{
\code{\link{pal_sgb}} fuer die SGB Standard-Farbpalette mit allen 5 Farben von \code{\link{pal_sgb_rot}};
\code{\link{pal_sgb_pref}} fuer eine SGB Farbpalette mit allen bevorzugten Farben;
\code{\link{seecol}} um die Farbpaletten anzusehen;
\code{\link{usecol}} um die Farbpaletten zu verwenden.

Other color palettes: 
\code{\link{pal_sgb_dunkelblau}},
\code{\link{pal_sgb_gelb}},
\code{\link{pal_sgb_grau}},
\code{\link{pal_sgb_hellblau}},
\code{\link{pal_sgb_pair}},
\code{\link{pal_sgb_pref}},
\code{\link{pal_sgb_signal}},
\code{\link{pal_sgb_web}},
\code{\link{pal_sgb}}
}
\concept{color palettes}
\keyword{datasets}
