% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_1.R
\docType{data}
\name{pal_sgb_signal}
\alias{pal_sgb_signal}
\title{Ampel SGB-Palette}
\format{
An object of class \code{data.frame} with 1 rows and 3 columns.
}
\usage{
pal_sgb_signal
}
\description{
\code{pal_sgb_signal} ist eine zusaetzliche SGB-Farbpalette als data frame mit 3 Farben
(Ampel- oder Signalfarben).
}
\details{
Die Farben sind wie bei einer Ampel angeordnet:
\enumerate{
\item top: rot oder "schlecht"
\item mit: gelb oder "Achtung"
\item unt: gruen oder "gut"
}
}
\examples{
pal_sgb_signal

pal_sgb_signal[2]    # (benannte) Farbe "signal2"
pal_sgb_signal[[2]]  # Farbe "signal2" oder "#EFDC60"

# Palette plotten:
seecol(pal_sgb_signal)

}
\seealso{
\code{\link{pal_sgb}} fuer die SGB Standard-Farbpalette mit allen 5 Farben von \code{\link{pal_sgb_rot}};
\code{\link{pal_sgb_pref}} fuer eine SGB Farbpalette mit allen bevorzugten Farben;
\code{\link{seecol}} um die Farbpaletten anzusehen;
\code{\link{usecol}} um die Farbpaletten zu verwenden.

Other color palettes: 
\code{\link{pal_sgb_dunkelblau}},
\code{\link{pal_sgb_gelb}},
\code{\link{pal_sgb_grau}},
\code{\link{pal_sgb_hellblau}},
\code{\link{pal_sgb_pair}},
\code{\link{pal_sgb_pref}},
\code{\link{pal_sgb_rot}},
\code{\link{pal_sgb_web}},
\code{\link{pal_sgb}}
}
\concept{color palettes}
\keyword{datasets}
