% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_1.R
\docType{data}
\name{pal_sgb_web}
\alias{pal_sgb_web}
\title{SGB Standard-Farbpalette (10 Farben, web)}
\format{
An object of class \code{data.frame} with 1 rows and 10 columns.
}
\usage{
pal_sgb_web
}
\description{
\code{pal_sgb_web} gibt die Standard-Farbpalette des SGB als data frame mit 10 Farben
}
\details{
Das ist die primaere SGB (web/sRGB) Palette.
}
\examples{
pal_sgb_web

# Auswahl nach Position:
pal_sgb_web[2]    # 2. Farbname "rot3" (als df)
pal_sgb_web[[2]]  # 2. Farbwert "#cb452c"

# Auswahl nach Name:
pal_sgb_web["rot3"]    # Farbe "rot3" (als df)
pal_sgb_web[["rot3"]]  # Farbwert "#cb452c"

# Palette plotten:
seecol(pal_sgb_web)

}
\seealso{
\code{\link{pal_sgb}} fuer die SGB Standard-Farbpalette mit allen 5 Farben von \code{\link{pal_sgb_rot}};
\code{\link{pal_sgb_pref}} fuer eine SGB Farbpalette mit allen bevorzugten Farben;
\code{\link{seecol}} um die Farbpaletten anzusehen;
\code{\link{usecol}} um die Farbpaletten zu verwenden.

Other color palettes: 
\code{\link{pal_sgb_dunkelblau}},
\code{\link{pal_sgb_gelb}},
\code{\link{pal_sgb_grau}},
\code{\link{pal_sgb_hellblau}},
\code{\link{pal_sgb_pair}},
\code{\link{pal_sgb_pref}},
\code{\link{pal_sgb_rot}},
\code{\link{pal_sgb_signal}},
\code{\link{pal_sgb}}
}
\concept{color palettes}
\keyword{datasets}
