% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_continuous}
\alias{scale_fill_continuous}
\title{Scale fill continuous with SGB-colors}
\usage{
scale_fill_continuous(...)
}
\arguments{
\item{...}{parameters passed on to \code{scale_fill_gradient()}.}
}
\description{
Overwrites the default continuous fill scale with SGB-colors. To use
another palette or custom colours, use \code{scale_fill_gradient()}.
}
\examples{
\dontrun{
ggplot(faithfuld, aes(waiting, eruptions)) +
  geom_raster(aes(fill = density))

# This gives the same:
ggplot(faithfuld, aes(waiting, eruptions)) +
  geom_raster(aes(fill = density)) +
  scale_fill_gradient()

# Custom gradient colors:
ggplot(faithfuld, aes(waiting, eruptions)) +
  geom_raster(aes(fill = density)) +
  scale_fill_gradient(low = pal_sgb_dunkelblau[5], high = pal_sgb_gelb[5])

# 3-color gradient (default):
ggplot(faithfuld, aes(waiting, eruptions)) +
  geom_raster(aes(fill = density)) +
  scale_fill_gradient2(midpoint = .02)

# 3-color gradient (custom):
ggplot(faithfuld, aes(waiting, eruptions)) +
  geom_raster(aes(fill = density)) +
  scale_fill_gradient2(low = pal_sgb_rot[5], mid = "white", high = pal_sgb_gelb[5],
                       midpoint = .02)

# n-colour gradient (default):
ggplot(faithfuld, aes(waiting, eruptions)) +
  geom_raster(aes(fill = density)) +
  scale_fill_gradientn()

# n-colour gradient (custom):
ggplot(faithfuld, aes(waiting, eruptions)) +
  geom_raster(aes(fill = density)) +
  scale_fill_gradientn(colors = pal_sgb_pref)

# Use `na.value = NA` to hide missing values but keep the original axis range
df_na <- data.frame(
  value = seq(1, 20),
  x = runif(20),
  y = runif(20),
  z1 = c(rep(NA, 10), rnorm(10)))
ggplot(df_na, aes(x = value, y)) +
  geom_bar(aes(fill = z1), stat = "identity") +
  scale_fill_gradient(na.value = NA)
}

}
