% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_discrete}
\alias{scale_fill_discrete}
\title{Scale fill discrete with SGB-colors}
\usage{
scale_fill_discrete(...)
}
\arguments{
\item{...}{parameters passed on to \code{scale_fill_manual()}.}
}
\description{
Overwrites the default discrete fill scale with SGB-palette \code{pal_sgb_pref}. To use
another palette or custom colours, use \code{scale_fill_manual()}.
}
\examples{
\dontrun{
ggplot(mpg, aes(as.factor(year), fill = drv)) +
  geom_bar(position = "dodge") +
    labs(x = "", y = "")

# This gives the same:
ggplot(mpg, aes(as.factor(year), fill = drv)) +
  geom_bar(position = "dodge") +
    labs(x = "", y = "") +
  scale_fill_discrete()

# To change fill scale:
ggplot(mpg, aes(as.factor(year), fill = drv)) +
  geom_bar(position = "dodge") +
    labs(x = "", y = "") +
 scale_fill_manual(values = rev(usecol(pal_sgb_dunkelblau)))
}

}
