% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_fun.R
\name{seecol}
\alias{seecol}
\alias{seepal}
\title{Farbpaletten plotten (um die Farben anzusehen).}
\usage{
seecol(
  pal = "sgb_all",
  n = "all",
  alpha = NA,
  hex = NULL,
  rgb = NULL,
  col_bg = NULL,
  col_brd = NULL,
  lwd_brd = NULL,
  grid = TRUE,
  title = NA,
  ...
)
}
\arguments{
\item{pal}{Eine Farbpalette (als Vektor mit Farben),
ein character string, die von seecol als keyword erkannt wurde, oder
mehrere Paletten als list.
Default: \code{pal = "sgb_all"}.

Erkannte keywords sind:

\enumerate{

  \item \code{"sgb_all"}: Alle Farbpaletten, die in \code{gsgb} definiert sind.

  \item \code{"sgb_basic"}: Alle SGB Standard-Farbpaletten.

  \item \code{"pair_all"}: Alle Farbpaletten mit paarweisen Farben.

  \item \code{"pref_all"}: Alle bevorzugten Farben und ihre Gradienten.

  \item \code{"grad_all"}:

}

\code{seecol} erkennt auch umgekehrte keywords (e.g., \code{"all_sgb"}) oder
keywords ohne \code{"sgb"} (z.B. \code{"basic"}).}

\item{n}{Anzahl Farben die angezeigt oder verwendet werden sollen.
Wenn \code{n} kleiner oder groesser als die Laenge der angegebenen Farbpalette \code{pal} ist,
wird die Farbalette reduziert oder extrapoliert
(mit \code{grDevices::colorRampPalette}).
Default: \code{n = "all"} (sprich zeige alle Farben der gewaehlten Palette).}

\item{alpha}{Ein Faktor, der die Deckkraft (opacity alpha) modifiziert
(wie in \code{\link{adjustcolor}}); typischerweise in [0,1].
Default: \code{NA} (d.h. keine aenderung der Deckkraft).}

\item{hex}{Sollen HEX Farbwerte gezeigt werden?
Default: \code{hex = NULL} (sprich zeige HEX Farbwerte
wenn es genug Platz hat um sie zu zeigen).}

\item{rgb}{Sollen RGB Farbwerte gezeigt werden?
Default: \code{rgb = NULL} (sprich zeige RGB Farbwerte
wenn es genug Platz hat um sie zu zeigen).}

\item{col_bg}{Hintergrundfarbe des Plots.
Default: \code{col_bg = NULL}.}

\item{col_brd}{Farbe der Umrandung (falls gezeigt).
Default: \code{col_brd = NULL}.}

\item{lwd_brd}{Linienstaerke der Umrandung (falls gezeigt).
Default: \code{lwd_brd = NULL}.}

\item{grid}{Raster im Plot verwenden?
Default: \code{grid = TRUE}.}

\item{title}{Plot-Titel
Default: \code{title = NA} generiert einen default Titel.}

\item{...}{Weitere grafische Parameter.
(passed to \code{plot}).}
}
\description{
\code{seecol} bietet eine Schnittstelle zum Plotten (oder "Sehen")
der Farben einer Palette oder fuer den Vergleich mehrerer Farbpaletten.
}
\details{
\code{seecol} hat 2 Hauptfunktionen, basierend auf den gegebenen \code{pal}-Argumenten:

\enumerate{

\item wenn \code{pal = "sgb_all"} (oder eine Liste mit mehreren Farbpaletten):

Plottet visuelle Vektoren aller aktuellen Farbpaletten, um sie zu vergleichen.

\item wenn mit \code{pal} eine bestimmte Farbpalette (oder ein Vektor mit mehreren Farben
oder Farbpaletten) gewaehlt wird:

Zeichnet die aktuelle Farbpalette und optionale Details zu ihren Farben.

}
}
\examples{
# Alle Farbpaletten anschauen:
seecol()  # gleich wie seecol(pal = "all")

# Details einer spezifischen Farbpalette anschauen:
seecol(pal_sgb_rot)

# Farben oder Farbpaletten kombinieren:
seecol(c(rev(pal_sgb_rot), pal_sgb_dunkelblau))  # Farbpalette kombinieren
seecol(c(rev(pal_sgb_hellblau), "white", pal_sgb_grau))  # Farbpalette und Farbnamen kombinieren
seecol(c("black", "blueviolet", "gold"))          # Farbnamen kombinieren

# Mit n die Farbpalette reduzieren oder erweitern:
seecol(n =  3)  # reduzierte Auswahl aller Paletten
seecol(n = 12)  # erweiterte Auswahl aller Paletten

seecol(pal_sgb, n = 5,
       title = "Reduzierte Version von pal_sgb (n = 5)")  # pal_sgb reduzieren
seecol(pal_sgb_dunkelblau, n = 8,
       title = "Erweiterte Version von pal_sgb_dunkelblau (n = 8)")  # pal_sgb_dunkelblau erweitern

# Farbpaletten erweitern und kombinieren
seecol(c(rev(pal_sgb_dunkelblau), "white", pal_sgb_gelb), n = 17,
       title = "Divergierende benutzerdefinierte Farbpalette mit 17 Farben")

# Benutzerdefinierte Farbpaletten definieren:
pal_mpg <- c("#007367", "white", "#D0D3D4")
names(pal_mpg) <- c("mpg green", "mpg white", "mpg grey")

# Erweiterte Farbpalette ansehen:
seecol(pal_mpg, n = 9, title = "Benutzerdefinierte Farbpalette fuer die Max-Planck-Gesellschaft")

# Farbpaletten vergleichen:
seecol(list(pal_mpg, pal_sgb_hellblau, pal_sgb), n = 5)

## Farbpaletten von anderen Packages ansehen:
# library(RColorBrewer)
# seecol(brewer.pal(name = "RdBu", n = 11))  # "RdBu" Palette von RColorBrewer

## Farbpalette erweitern
# seecol(brewer.pal(name = "RdBu", n = 11), n = 15)  # Farbpalette auf 15 Farben erweitern

}
\seealso{
\code{\link{usecol}} um eine Farbpalette zu verwenden;
\code{\link{pal_sgb}} die SGB Standard-Farbpalette.

Other Farbfunktionen: 
\code{\link{newpal}()},
\code{\link{usecol}()}
}
\concept{Farbfunktionen}
