% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_1.R
\docType{data}
\name{SGBdunkelblau}
\alias{SGBdunkelblau}
\title{Dunkelblaue SGB-Farbe}
\format{
An object of class \code{character} of length 1.
}
\usage{
SGBdunkelblau
}
\description{
\code{SGBdunkelblau} gibt die bevorzugte Farbe von \code{\link{pal_sgb_dunkelblau}}
(als atomic HEX character Wert) und ist definiert als \code{\link{pal_sgb_dunkelblau}[[5]]}.
}
\examples{
SGBdunkelblau  # HEX character "#002060" (als Wert)
all.equal(SGBdunkelblau, pal_sgb_dunkelblau[[5]])  # TRUE (gleiche HEX Werte)

seecol(SGBdunkelblau)  # Farbe und Details ansehen

}
\seealso{
\code{\link{pal_sgb_dunkelblau}} fuer die entsprechende Farbpalette;
\code{\link{pal_sgb}} fuer die SGB Standard-Farbpalette mit allen 5 Farben von \code{\link{pal_sgb_rot}};
\code{\link{pal_sgb_pref}} fuer eine SGB Farbpalette mit allen bevorzugten Farben;
\code{\link{seecol}} um die Farbpaletten anzusehen;
\code{\link{usecol}} um die Farbpaletten zu verwenden.

Other preferred colors: 
\code{\link{SGBgelb}},
\code{\link{SGBgrau}},
\code{\link{SGBhellblau}},
\code{\link{SGBrot}}
}
\concept{preferred colors}
\keyword{datasets}
