% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def.R
\docType{data}
\name{SGBgelb}
\alias{SGBgelb}
\title{Gelbe SGB-Farbe}
\format{
An object of class \code{character} of length 1.
}
\usage{
SGBgelb
}
\description{
\code{SGBgelb} gibt die bevorzugte Farbe von \code{\link{pal_sgb_gelb}}
(als atomic HEX character Wert) und ist definiert als \code{\link{pal_sgb_gelb}[[5]]}.
}
\examples{
SGBgelb  # HEX character "#ffc000" (als Wert)
all.equal(SGBgelb, pal_sgb_gelb[[5]])  # TRUE (gleiche HEX Werte)

seecol(SGBgelb)  # Farbe und Details ansehen

}
\seealso{
\code{\link{pal_sgb_gelb}} fuer die entsprechende Farbpalette;
\code{\link{pal_sgb}} fuer die SGB Standard-Farbpalette mit allen 5 Farben von \code{\link{pal_sgb_rot}};
\code{\link{pal_sgb_pref}} fuer eine SGB Farbpalette mit allen bevorzugten Farben;
\code{\link{seecol}} um die Farbpaletten anzusehen;
\code{\link{usecol}} um die Farbpaletten zu verwenden.

Other preferred colors: 
\code{\link{SGBdunkelblau}},
\code{\link{SGBgrau}},
\code{\link{SGBhellblau}},
\code{\link{SGBrot}}
}
\concept{preferred colors}
\keyword{datasets}
