% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_1.R
\docType{data}
\name{SGBgrau}
\alias{SGBgrau}
\title{Graue SGB-Farbe}
\format{
An object of class \code{character} of length 1.
}
\usage{
SGBgrau
}
\description{
\code{SGBgrau} gibt die bevorzugte Farbe von \code{\link{pal_sgb_grau}}
(als atomic HEX character Wert) und ist definiert als \code{\link{pal_sgb_grau}[[3]]}.
}
\examples{
SGBgrau  # HEX character "#7c7c7c" (als Wert)
all.equal(SGBgrau, pal_sgb_grau[[3]])  # TRUE (gleiche HEX Werte)

seecol(SGBgrau)  # Farbe und Details ansehen

}
\seealso{
\code{\link{pal_sgb_grau}} fuer die entsprechende Farbpalette;
\code{\link{pal_sgb}} fuer die SGB Standard-Farbpalette mit allen 5 Farben von \code{\link{pal_sgb_rot}};
\code{\link{pal_sgb_pref}} fuer eine SGB Farbpalette mit allen bevorzugten Farben;
\code{\link{seecol}} um die Farbpaletten anzusehen;
\code{\link{usecol}} um die Farbpaletten zu verwenden.

Other preferred colors: 
\code{\link{SGBdunkelblau}},
\code{\link{SGBgelb}},
\code{\link{SGBhellblau}},
\code{\link{SGBrot}}
}
\concept{preferred colors}
\keyword{datasets}
