% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_1.R
\docType{data}
\name{SGBrot}
\alias{SGBrot}
\title{Rote SGB-Farbe}
\format{
An object of class \code{character} of length 1.
}
\usage{
SGBrot
}
\description{
\code{SGBrot} gibt die bevorzugte Farbe von \code{\link{pal_sgb_rot}}
(als atomic HEX character Wert) und ist definiert als \code{\link{pal_sgb_rot}[[4]]}.
}
\examples{
SGBrot  # HEX character "#990000" (als Wert)
all.equal(SGBrot, pal_sgb_rot[[4]])  # TRUE (gleiche HEX Werte)

seecol(SGBrot)  # Farbe und Details ansehen

}
\seealso{
\code{\link{pal_sgb_rot}} fuer die entsprechende Farbpalette;
\code{\link{pal_sgb}} fuer die SGB Standard-Farbpalette mit allen 5 Farben von \code{\link{pal_sgb_rot}};
\code{\link{pal_sgb_pref}} fuer eine SGB Farbpalette mit allen bevorzugten Farben;
\code{\link{seecol}} um die Farbpaletten anzusehen;
\code{\link{usecol}} um die Farbpaletten zu verwenden.

Other preferred colors: 
\code{\link{SGBdunkelblau}},
\code{\link{SGBgelb}},
\code{\link{SGBgrau}},
\code{\link{SGBhellblau}}
}
\concept{preferred colors}
\keyword{datasets}
