% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_fun.R
\name{usecol}
\alias{usecol}
\title{Benutze eine Farbpalette}
\usage{
usecol(
  pal = pal_sgb,
  n = "all",
  alpha = NA,
  use_names = FALSE,
  use_col_ramp = FALSE
)
}
\arguments{
\item{pal}{Eine Farbpalette (als Vektor mit Farben oder eine Farbpalette).
Default: \code{pal = \link{pal_sgb}}.}

\item{n}{Ein ganzzahliger Wert, der die gewuenschte Anzahl von Farben aus der Palette angibt.
Fuer alle Paletten, die in \code{gsgb} definiert sind, verwendet es standardmaessig eine vordefinierte
Auswahl von Farben, wenn die gewuenschte Anzahl von Farben kleiner als die verfuegbare Anzahl ist.
Fuer alle anderen Paletten und \code{n} groesser als \code{length(pal)} erweitert es die Palette mit
\code{\link{colorRampPalette}}.}

\item{alpha}{Ein Faktor, der die Deckkraft (opacity alpha) modifiziert
(wie in \code{\link{adjustcolor}}); typischerweise in [0,1].
Default: \code{NA} (d.h. keine aenderung der Deckkraft).}

\item{use_names}{Ein logischer Wert der definiert, ob Farben als benannter Vektor zurueckgegeben werden
sollen. (Default ist \code{FALSE}, zur Kompatibilitaet mit \code{ggplot}).}

\item{use_col_ramp}{Ein logischer Wert, der angibt, ob der Default der vorausgewaehlten Farben
ueberschrieben werden soll und stattdessen \code{\link{colorRampPalette}} zur Verarbeitung von
\code{n} verwendet werden soll.}
}
\description{
\code{usecol} erlaubt die Verwendung einer Farbpalette \code{pal} (z.B. zum plotten).
}
\examples{
usecol(pal = pal_sgb, n = "all")  # default Farbpalette
usecol(pal = pal_sgb, n =  4)     # Auswahl von n vorausgewaehlten Farben
usecol(pal = pal_sgb, n = 20)     # Farbpalette erweitern

# Neue Farbpalette mischen:
pal_1 <- usecol(pal = c(rev(pal_sgb_rot), "white", pal_sgb_gelb))
seecol(pal_1)

# Farbpalette mischen und erweitern:
pal_2 <- usecol(pal = c(rev(pal_sgb_dunkelblau), "white", pal_sgb_hellblau), n = 20)
seecol(pal_2)

# Eine eigene Farbpalette definieren und verwenden:
pal_ahv13 <- c("#E94C25", "white", "black")
names(pal_ahv13) <- c("orange_ahv", "white", "black")

pal_3 <- usecol(pal_ahv13, n = 7)
seecol(pal_3)

}
\seealso{
\code{\link{seecol}} um Farbpaletten zu plotten;
\code{\link{pal_sgb}} fuer die SGB Standard-Farbpalette

Other Farbfunktionen: 
\code{\link{newpal}()},
\code{\link{seecol}()}
}
\concept{Farbfunktionen}
