% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_discrete}
\alias{scale_colour_discrete}
\title{Default scale mit SGB-Farben}
\usage{
scale_colour_discrete(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Default scale mit SGB-Farben
}
