% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_discrete}
\alias{scale_fill_discrete}
\title{Default scale mit SGB-Farben}
\usage{
scale_fill_discrete(...)
}
\arguments{
\item{...}{Beliebige Parameter die an die Funktion \code{scale_fill_manual()} weitergegeben werden koennen.}
}
\value{
see return of \code{scale_fill_manual()}.
}
\description{
Default scale mit SGB-Farben
}
